ptr ref change

+++ ccip_mmio/sw/main.cpp
@@ -234,20 +234,20 @@ int main(int argc, char *argv[]) {
 	// Start time
 	gettimeofday(&start, nullptr);
 
-	for (int i = 0; i <DIM; i += 8){
-		for (int j = 0; j < DIM; j+=8){
-			for (int ii = 0; ii < 8; ii ++){
+	for (ptrdiff_t i = 0; i <DIM; i += 8){
+		for (ptrdiff_t j = 0; j < DIM; j+=8){
+			for (ptrdiff_t ii = 0; ii < 8; ii ++){
 				send_row_C(ii, output[i+ii][j],afu);
 			}
 
-			for (int k = 0; k <DIM; k += 8){
-				for (int ii = 0; ii < 8; ii ++){
+			for (ptrdiff_t k = 0; k <DIM; k += 8){
+				for (ptrdiff_t ii = 0; ii < 8; ii ++){
 					send_row_A(ii, A_vals[i+ii][k],afu);
 					send_row_B(ii, B_vals[i+ii][j],afu);
 				}	
 				afu.write(0x0400, 100);
 			}
-			for (int ii = 0; ii < 8; ii ++){
+			for (ptrdiff_t ii = 0; ii < 8; ii ++){
 				unpack_from_C(ii, output[i+ii][j],afu);
 			}
